# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Management System API
  version: 1.0.0
  description: |
    API documentation for the Medical Management System backend.

    This API provides endpoints for managing users, doctors, patients, appointments, and medical records.
    Authentication is required for most endpoints using JWT tokens.
  contact:
    name: Medical Management System
    email: support@medical-system.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.medical-system.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /appointments:
    post:
      tags:
        - Appointments
      summary: Create a new appointment (patient only)
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointment'
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Time slot unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Appointments
      summary: Get appointments (filtered by user role)
      operationId: getAppointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - scheduled
              - confirmed
              - cancelled
              - completed
              - no_show
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: doctor_id
          in: query
          schema:
            type: integer
        - name: patient_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
...
  /appointments/{id}:
    get:
      tags:
        - Appointments
      summary: Get appointment by ID
      operationId: getAppointmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Appointments
      summary: Update appointment
      operationId: updateAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointment'
      responses:
        '200':
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
  /appointments/{id}/cancel:
    put:
      tags:
        - Appointments
      summary: Cancel appointment
      operationId: cancelAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'

  # Medical Records endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum:
            - admin
            - doctor
            - patient
          example: patient
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        phone:
          type: string
          example: '+1234567890'
        date_of_birth:
          type: string
          format: date
          example: '1990-01-01'
        gender:
          type: string
          enum:
            - male
            - female
            - other
          example: male
        address:
          type: string
          example: 123 Main St, City, State 12345
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        license_number:
          type: string
          example: MD123456
        specialization:
          type: string
          example: Cardiology
        experience_years:
          type: integer
          example: 10
        consultation_fee:
          type: number
          format: decimal
          example: 150
        education:
          type: string
          example: MD from Harvard Medical School
        bio:
          type: string
          example: Experienced cardiologist with 10 years of practice
        is_available:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        patient_id:
          type: integer
          example: 1
        doctor_id:
          type: integer
          example: 1
        appointment_date:
          type: string
          format: date
          example: '2023-12-25'
        appointment_time:
          type: string
          format: time
          example: '10:00'
        duration:
          type: integer
          example: 30
        status:
          type: string
          enum:
            - scheduled
            - confirmed
            - cancelled
            - completed
            - no_show
          example: scheduled
        appointment_type:
          type: string
          enum:
            - consultation
            - follow_up
            - emergency
            - routine_checkup
          example: consultation
        reason:
          type: string
          example: Regular checkup
        notes:
          type: string
          example: Patient has been feeling chest pain
        reminder_sent:
          type: boolean
          example: false
        cancelled_by:
          type: integer
          example: 1
        cancelled_at:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        cancellation_reason:
          type: string
          example: Patient unavailable
        patient:
          $ref: '#/components/schemas/User'
        doctor:
          $ref: '#/components/schemas/Doctor'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
    CreateAppointment:
      type: object
      required:
        - doctor_id
        - appointment_date
        - appointment_time
        - appointment_type
      properties:
        doctor_id:
          type: integer
          example: 1
        appointment_date:
          type: string
          format: date
          example: '2023-12-25'
        appointment_time:
          type: string
          format: time
          example: '10:00'
        duration:
          type: integer
          example: 30
        appointment_type:
          type: string
          enum:
            - consultation
            - follow_up
            - emergency
            - routine_checkup
          example: consultation
        reason:
          type: string
          example: Regular checkup
        notes:
          type: string
          example: Patient has been feeling chest pain
    UpdateAppointment:
      type: object
      properties:
        appointment_date:
          type: string
          format: date
          example: '2023-12-25'
        appointment_time:
          type: string
          format: time
          example: '10:00'
        duration:
          type: integer
          example: 30
        status:
          type: string
          enum:
            - scheduled
            - confirmed
            - cancelled
            - completed
            - no_show
          example: confirmed
        appointment_type:
          type: string
          enum:
            - consultation
            - follow_up
            - emergency
            - routine_checkup
          example: consultation
        reason:
          type: string
          example: Regular checkup
        notes:
          type: string
          example: Patient has been feeling chest pain
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid input provided
        message:
          type: string
          example: The request could not be processed
        success:
          type: boolean
          example: false
...